requirements.txt
streamlit>=1.30.0
qrcode[pil]
Pillow
reportlab
streamlit-drawable-canvas
pymupdf
openpyxl
supabase>=2.0.0
